Revert convert_filename from shared as no longer needed to be shared
